reverted refactoring for fetchBucketAndPath and enableGcsLifecycle
